Extract services/praxis_duel.py so praxis.py imports the duel surface at module level
Closes the five function-body `from services.duel import` sites in
services/praxis.py (and the sixth in services/vote.py) by moving the three
praxis-id-keyed duel functions into a leaf module both sides can import at
module level, per SPEC-backend-architecture section 11 and the
services/duel_outcome.py precedent (#1391, #2872). Pure move.
