+++ backend/tests/unit/ruff_allowlist.py
@@ -70,7 +70,7 @@ from __future__ import annotations
 
 #: The headline number. Asserted against the sum of the entries below, so it
 #: cannot drift into being a separate claim about the code.
-RUFF_FINDING_TOTAL = 638
+RUFF_FINDING_TOTAL = 637
 
 #: ``<path relative to backend/>::<rule code>`` -> how many that file is still
 #: allowed. Sorted by path; keep it sorted so diffs stay readable.
@@ -164,7 +164,7 @@ RUFF_ALLOWLIST: dict[str, int] = {
     "services/character_stats.py::E501": 4,
     "services/character_stats.py::I001": 1,
     "services/collab_consensus.py::E501": 3,
-    "services/duel.py::E501": 5,
+    "services/duel.py::E501": 4,
     "services/duel.py::I001": 1,
     "services/era.py::E501": 1,
     "services/era.py::I001": 1,
